# swagger.yaml
openapi: 3.0.0
info:
  title: VOLCANOE  API DOC
  description: An Express Volcanoe API
  version: 1.0.0
servers:
  - url: http://myapi.us-east-1.elasticbeanstalk.com/
    description: Production server
  - url: http://localhost:3000/
    description:  Development server 
components:
  securitySchemes:
    ApiTokenss:        # arbitrary name for the security scheme
          
      type: http
      scheme: bearer
    
    ApiKey:        # arbitrary name for the security scheme
      type: apiKey
      in: header       # can be "header", "query" or "cookie"
      name: apikey
  schemas:
...
    Volcanoe:
      type: object
      required:
        - Region
        - Number
        - Volcanoe_Name
        - Country
        - Location
        - Latitude
        - Longitude
        - Elevation
        - Type
        - Status
      properties:
        Region:
          type: string
          description: The volcanoe region
        Number:
          type: string
          description: The volcanoe number
        Volcanoe_Name:
          type: string
          description: The volcanoe name
        Country:
          type: string
          description: Country the volcanoe Occured
        Location:
          type: string
          description: The location of the volcanoe
        Latitude:
            type: string
            description: The Latitude of the volcanoe
        Longitude:
            type: string
            description: The Longitude of the volcanoe
        Elevation:
            type: string
            description: The volcanoe Elevation
        Type:
            type: string
            description: The type of volcanoe
        Status:
            type: string
            description: The volcanoe status
      example:
        Region: West Indies
        Number: 1600-21-
        Volcanoe_Name: Piparo
        Country: Trinidad
        Location: Trinidad
        Latitude: 10
        Longitude: -61
        Elevation: 140
        Type: Mud Volcanoe
        Status: Historical
...
paths:
...
  /volcanoes/{id}:
    get:
      summary: Get the volcanoe by id
      tags: [Volcanoes]
      parameters:
        - in: path
          name: id
          schema:
            type: string
          required: true
          description: The volcanoe id
      responses:
        '200':
          description: Get a volcanoe by id successful
          contens:
            application/json:
              schema:
                $ref: '#/components/schemas/Volcanoe'
        '404':
          description: The volcanoe was not found
...
  /volcanoes:
    get:
      summary: Returns the list of all the volcanoes
      tags: [Volcanoes]
      responses:
        '200':
          description: The list of all the volcanoes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Volcanoe'
        '404':
          description: Not found